Update xliff-example1-0.9.3.json to show whitespace
The Xliff 2.0 example [here](https://github.com/oasis-tcs/xliff-omos-jliff/blob/master/JLIFF-schema-draft/xliff-example1.xml#L11) shows whitespace around the source and target text elements aaa and AAA that are not present in the Jliff draft example. Whitespace in xml is particular pain point for some serializers so it would be good to show how the spaces are represented in Jliff.