val, mtc and ctr JLIFF spec draft update
This is update partly covers mtc and ctr. There are internal schema structuring/naming considerations to be made by the TC to improve the mapping of module attributes and elements to JLIFF, mostly related to the anonymity of objects and arrays containing objects (i.e. such desisions will not impact JLIFF JSON docs).